pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tapdata-sdk"
version = "0.2.0"
description = "A Python client library for interacting with Tapdata API"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
keywords = ["tapdata", "api", "client", "sdk", "data-integration"]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "requests>=2.25.0",
    "pycryptodome>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "types-requests>=2.25.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/tapdata-sdk"
Documentation = "https://github.com/yourusername/tapdata-sdk#readme"
Repository = "https://github.com/yourusername/tapdata-sdk"
Issues = "https://github.com/yourusername/tapdata-sdk/issues"

[tool.setuptools]
packages = ["tapdata_sdk"]

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=tapdata_sdk --cov-report=html --cov-report=term"
